requirements.txt
pandas==1.5.3
openpyxl==3.1.0
matplotlib==3.6.3
scikit-learn==1.2.1
numpy==1.24.3
torch==1.13.1
tabulate==0.9.0
seaborn==0.12.2
dython==0.5.1
tqdm==4.64.1
hydra-core==1.3.1
optuna==3.1.0
catboost==1.2.1
rdt==1.3.0
synthcity==0.2.6
pydantic==1.10.8
zero==0.9.1
libzero==0.0.8
icecream==2.1.3
tomli_w=1.0.0
rtdl==0.0.13